Fix file locations
