+++ model-worker.ts
